change type

+++ include/cuda_pointclouds.h
@@ -5,6 +5,6 @@
 #include <pcl/point_types.h>
 
 __global__ void setPointLabelKernel(pcl::PointXYZRGBL *points, std::uint32_t label, int num_points);
-__global__ void transformPointKernel(pcl::PointXYZRGBL *points, Eigen::Matrix4f transform, int num_points);
+__global__ void transformPointKernel(pcl::PointXYZRGBL *points, Eigen::Matrix4d transform, int num_points);
 
 #endif
